+++ docs/changelog.md
@@ -9,7 +9,7 @@
 
 ### Changes
 * Added TypeScript case to regenerate-protobuf-bindings.sh so that now it considers TypeScript within the shell script
-* Generated TypeScript bindings which can now be ouputted by the wrapper script which calls down to developer-tools
+* Generated TypeScript bindings which can now be outputted by the wrapper script which calls down to developer-tools
 
 # 0.2.0
 ### Changes