+++ docs/changelog.md
@@ -1,4 +1,7 @@
 # TBD
+### Changes
+* Updated TypeScript binding file
+* Generated JavaScript binding files which were added  after calling down to developer-tools
 
 # 0.2.1
 ### Features
@@ -6,7 +9,7 @@
 
 ### Changes
 * Added TypeScript case to regenerate-protobuf-bindings.sh so that now it considers TypeScript within the shell script
-* Generated TypeScript bindings which can now be ouputted by the wrapper script which calls down to developer-tools
+* Generated TypeScript bindings which can now be outputted by the wrapper script which calls down to developer-tools
 
 # 0.2.0
 ### Changes