+++ docs/changelog.md
@@ -1,6 +1,13 @@
 # TBD
+### Changes
+* Updated TypeScript binding file
+* Generated JavaScript binding files which were added  after calling down to developer-tools
+
 ### Features
 * Added an `ExecuteBulkCommands` endpoint to the API that accepts JSON-serialized bulk command information and executes it against the API container
+
+### Fixes
+* Don't panic when a user passes in a null partition connections map to `NetworkContxt.repartitionNetwork`
 
 # 0.2.1
 ### Features
@@ -8,7 +15,7 @@
 
 ### Changes
 * Added TypeScript case to regenerate-protobuf-bindings.sh so that now it considers TypeScript within the shell script
-* Generated TypeScript bindings which can now be ouputted by the wrapper script which calls down to developer-tools
+* Generated TypeScript bindings which can now be outputted by the wrapper script which calls down to developer-tools
 
 # 0.2.0
 ### Changes